backend/requirements.txt
Flask==3.1.0
Faker==0.7.4
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
PyMySQL==1.1.1
pdbpp==0.10.3
python-dotenv==1.1.0
passlib==1.7.4